Mortgage Page Updated and Linked Up

+++ PalindromeChecker/PalindromeChecker/Models/LoanModel.cs
@@ -13,7 +13,7 @@ namespace PalindromeChecker.Models
         public decimal LoanAmount { get; set; }
         public decimal TotalInterest { get; set; }
         public decimal TotalCost { get; set; }
-        public int Term { get; set; }
+        public decimal Term { get; set; }
         public string Message { get; set; }
 
         public List<LoanPayment> Payments { get; set; } = new List<LoanPayment>();
